Adding a like box to gh pages.
Thanks to the facebook. Now the old versions are deprecated and the new
vesrion is not compatible with older html vesions. I tried to find a
iframe solution. Iframe like box are still functional but i was not able
to do so, so my final solution was to change the iframe to object
solution here:

http://blog.chapagain.com.np/w3c-validation-iframe-error-with-xhtml-1-0-strict-doctype/